students/KolodichMaksim/task_02/src/requirements.txt
Flask>=2.2
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
Flask-JWT-Extended>=4.4
Flask-Cors>=4.0
Flask-Talisman>=1.0
python-dotenv>=1.0
